% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adicionar_grupo_tabela.R
\name{adicionar_grupo_tabela}
\alias{adicionar_grupo_tabela}
\title{Insere grupo em dataframe respeitando os cd_processos}
\usage{
adicionar_grupo_tabela(df = NULL, by = NULL, g = NULL)
}
\arguments{
\item{df}{Dataframe}

\item{by}{colunas: processo, cd_processo_pg etc}

\item{g}{Inteiro}
}
\value{
Mesmo dataframe com a coluna informando o grupo
   ao qual pertence o processo ou outra coluna.
}
\description{
Insere grupo em dataframe respeitando os cd_processos
}
