% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{azure_ai_mensagens}
\alias{azure_ai_mensagens}
\title{Auxilia na construção de mensagens}
\usage{
azure_ai_mensagens(x, tipo_texto, destaques, perguntas, colunas)
}
\arguments{
\item{x}{decisão judicial}

\item{tipo_texto}{tipo de texto}

\item{destaques}{Vetor de palavras ou expressões a destacar}

\item{perguntas}{Vetor de perguntas}

\item{colunas}{Vetor de colunas}
}
\value{
Lista de mensagens
}
\description{
Auxilia na construção de mensagens
}
