% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{azure_openai_extrair}
\alias{azure_openai_extrair}
\title{Obtêm resposta da Azure Openai para peças processuais.}
\usage{
azure_openai_extrair(
  x,
  tipo_texto = "decisão",
  perguntas,
  colunas,
  destaques = NULL,
  recurso,
  implementacao,
  versao_api = "2023-10-01-preview",
  api_key = NULL,
  temperatura = 0,
  max_tokens = 4000,
  presence_penalty = 0,
  frequency_penalty = 0,
  logit_bias = NULL
)
}
\arguments{
\item{x}{Vetor de documentos,e.g., julgados, petições iniciais.}

\item{tipo_texto}{c("decisão", "acórdão","petição",...). Informe o tipo
de texto a ser analisado.}

\item{perguntas}{Vetor de perguntas.}

\item{colunas}{Colunas/chaves das respostas em json}

\item{destaques}{Palavras ou expressões relevantes.}

\item{recurso}{Nome do recurso ou endpoint. Nome do recurso criado por você. 
Você pode informar tanto o recuso quanto o endpoint que contêm o recurso.
Para encontrá-lo, vá para Azure OpenAI Studio > Playground > View code.}

\item{implementacao}{OU deployment id. Este valor corresponde ao nome
dado por você ao implementar (deploy) um modelo. Pode ser encontrado em
Resource Management > Model Deployments.}

\item{versao_api}{Versão no formato: yyyy-mm-dd com ou sem preview.}

\item{api_key}{Chave. Você pode armazená-la na variável de ambiente
AZURE_OPENAI_KEY.}

\item{temperatura}{Nível de aleatoriedade. Padrão: 0, ou seja, determinístico.}

\item{max_tokens}{Máximo de tokens na resposta.}

\item{presence_penalty}{Valores entre -2 e 2. Quanto maior o valor
mais conservador (menor o número) na geração de palavras ou frases repetidas.}

\item{frequency_penalty}{Valores entre -2 e 2. Valores positivos aumentam a
variabilidade de palavras.}

\item{logit_bias}{Altera a chance de um token aparecer na resposta.}
}
\value{
json
}
\description{
Obtêm resposta da Azure Openai para peças processuais.
}
