% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_gender.R
\name{br_gender}
\alias{br_gender}
\title{Atribui sexo a nomes com base no primeiro nome a partir da base do IBGE.}
\usage{
br_gender(nomes)
}
\arguments{
\item{nomes}{Vetor com os nomes completos ou não para realizar a busca.}
}
\value{
tibble com cinco colunas: primeiro_nome,f para feminino, m para masculino,sexo e nomes=para o nome
}
\description{
Atribui sexo a nomes com base no primeiro nome a partir da base do IBGE.
}
