% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busca_fuzzy.R
\name{busca_fuzzy}
\alias{busca_fuzzy}
\title{Procura a palavra ou frase do segundo vetor que melhor 
   se aproxima do primeiro. Particularmente útil para 
   comparar nomes de municípios.}
\usage{
busca_fuzzy(x, y)
}
\arguments{
\item{x}{Vetor de strings de referência.}

\item{y}{Vetor de strings a serem buscados.}
}
\value{
vetor com as strings de y próximos
    de x.
}
\description{
Procura a palavra ou frase do segundo vetor que melhor 
   se aproxima do primeiro. Particularmente útil para 
   comparar nomes de municípios.
}
