% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jus_kwic.R
\name{chop_collapse}
\alias{chop_collapse}
\title{junta trechos e colapsa em um único texto}
\usage{
chop_collapse(df, coluna)
}
\arguments{
\item{df}{Data.frame}

\item{coluna}{coluna}
}
\value{
Mesmo data.frame com coluna colapsada
}
\description{
junta trechos e colapsa em um único texto
}
