% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificar_antecedentes.R
\name{classifica_antecedentes}
\alias{classifica_antecedentes}
\title{Classifica antecedência, primariedade e reincidência}
\usage{
classifica_antecedentes(x)
}
\arguments{
\item{x}{Vetor de julgados ou trechos
extraídos com jus_kwic}
}
\value{
Vetor com classificações
}
\description{
Classifica antecedência, primariedade e reincidência
}
\examples{
julgados <-
c(`1` = "Possui maus antecedentes. Julgado procedente",
 `2` = "Fulano não é primário.
     Condeno Fulano a 10 anos de reclusão.
     Deve recorrer em liberdade")

}
