% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_appeal.R
\name{classify_appeal}
\alias{classify_appeal}
\title{Classify high courts opinions by decision}
\usage{
classify_appeal(x, text, decision)
}
\arguments{
\item{x}{data.frame with a column classified}

\item{text}{Column where the "dispositivo" can be found.}

\item{decision}{Column name to be added with the decision.}
}
\value{
same data.frame with a new column called decision.
}
\description{
Classify high courts opinions by decision
}
