% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearPt.R
\name{clearPt}
\alias{clearPt}
\title{Function clearPt}
\usage{
clearPt(string, lower = T, stopwords = T, accent = T, punctuation = T,
  whitespace = T, numbers = T, ordinal = T)
}
\arguments{
\item{string}{object to be submitted to conversion and removals.}

\item{lower}{logical Convert to lower case.}

\item{stopwords}{logial Removes portuguese-BR stopwords.}

\item{accent}{logial remove all diacritics}

\item{punctuation}{logical Romoves punctuation.}

\item{whitespace}{logical Remove whitespaces}

\item{numbers}{logical Removes numbers}

\item{ordinal}{logical Removes ordinal superscripts}
}
\description{
Optionally Removes stopwords, numbers, ordinals, whitespaces, punctuation, accents, and convert to lower case.
}
\examples{
\dontrun{
clearPt(texto)
}
}
