% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnj_classe.R
\name{cnj_classe}
\alias{cnj_classe}
\title{Busca códigos de classes processuais no CNJ}
\usage{
cnj_classe(texto = NULL)
}
\arguments{
\item{texto}{Termos a serem procurados}
}
\value{
Tibble com códigos e classes
}
\description{
Busca códigos de classes processuais no CNJ
}
\examples{
\dontrun{
df <- cnj_classe("procedimento comum")
}
}
