% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnj_ultimo.R
\name{cnj_ultimo}
\alias{cnj_ultimo}
\title{Encontra a quantidade de processos distribuídos numa unidade judiciária.}
\usage{
cnj_ultimo(x, ano, nivel, uf, distribuidor, funcao)
}
\arguments{
\item{x}{Qualquer inteiro que seguramente é superior ao número de processos distribuídos}

\item{ano}{Indicar o ano em questão}

\item{nivel}{nível do judiciário}

\item{uf}{unidade federativa.}

\item{distribuidor}{código do distribuídor. Encontrar no data-raw.}

\item{funcao}{Função a ser aplicada para baixar: baixar_cpopg, baixar_processo, etc.}
}
\value{
Quantidade máxima aproximada de processos distribuídos. Pode haver um erro, para cima,
    de cinco processos.
}
\description{
Encontra a quantidade de processos distribuídos numa unidade judiciária.
}
