% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colapsar_coluna.R
\name{colapsar_coluna}
\alias{colapsar_coluna}
\title{Colapsar coluna}
\usage{
colapsar_coluna(df, grupo, coluna, collapse = "\\n")
}
\arguments{
\item{df}{data.frame}

\item{grupo}{grupo}

\item{coluna}{coluna a ser colapsada}

\item{collapse}{separador}
}
\value{
mesmo data.frame com a coluna colapsada
}
\description{
Colapsar coluna
}
