% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colapsar_partes.R
\name{colapsar_partes}
\alias{colapsar_partes}
\title{Colapsar as partes para visualização com DT}
\usage{
colapsar_partes(df, grupo, ..., collapse = "<br>")
}
\arguments{
\item{df}{Data.frame}

\item{grupo}{Coluna agrupadora, geralmente processo}

\item{...}{Colunas a serem aninhadas e colapsadas}

\item{collapse}{separador dos resultados. Coloquei
"<br>" pq uso no datatable.}
}
\value{
Mesmo def com valores aninhados e colapsados
}
\description{
Colapsar as partes para visualização com DT
}
