% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_form.R
\name{create_form}
\alias{create_form}
\title{Creates list from parsed form data.}
\usage{
create_form(string)
}
\arguments{
\item{string}{parsed form data.}
}
\value{
list with named form's elements
}
\description{
Creates list from parsed form data.
}
\details{
It only works with unnested form data
}
\examples{
string <- "visaoId=tjdf.djeletronico.comum.internet.apresentacao.VisaoDiarioEletronicoConsultaLivre&controladorId=tjdf.djeletronico.comum.internet.apresentacao.ControladorDiarioEletronicoConsultaLivre&idDoUsuarioDaSessao=&nomeDaPagina=montagem&comando=consultar&enderecoDoServlet=djeletronico&visaoAnterior=tjdf.djeletronico.comum.internet.apresentacao.VisaoDiarioEletronicoConsultaLivre&skin=&tokenDePaginacao=2&idJanelaAbrirAjax=&idJanelaFecharAjax=&idJanelaAbrirIsModalAjax=false&internet=1&textoDaConsulta=0000011-53.2016.8.07.0016&data=08\%2F12\%2F2018"
body <- create_form(string)

}
