% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_arquivo.R
\name{cria_arquivo}
\alias{cria_arquivo}
\title{Cria nome de arquivo a ser baixado}
\usage{
cria_arquivo(
  diretorio = ".",
  prefixo = "",
  sufixo = "",
  hora = TRUE,
  ext = ".html"
)
}
\arguments{
\item{diretorio}{Diretório}

\item{prefixo}{Identificador inicial}

\item{sufixo}{Identificador final}

\item{hora}{Incluir a hora da coleta?}

\item{ext}{Informar a extensão. Padrão é .html}
}
\value{
Nome do arquivo
}
\description{
Cria nome de arquivo a ser baixado
}
\examples{
cria_arquivo(diretorio = "data-raw/cjpg", prefixo = "cjpg_")
}
