% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criar_orgaos.R
\name{criar_orgaos}
\alias{criar_orgaos}
\title{Cria sequência ordinal de órgão julgador}
\usage{
criar_orgaos(nome = NULL, numero = NULL)
}
\arguments{
\item{nome}{ex. vara, câmara, turma}

\item{numero}{quantidade}
}
\value{
vetor de órgãos
}
\description{
Cria sequência ordinal de órgão julgador
}
