% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_sg.R
\name{decision_sg}
\alias{decision_sg}
\title{Label high courts decisions as in favor (yes) or against (not) the appelant.}
\usage{
decision_sg(docs, word, valuetype)
}
\arguments{
\item{docs}{with high court decision.}

\item{word}{word to be searched. Usualy its "provimento" for appeals or "concedido|denegado"
for writ.}

\item{valuetype}{how to interpret keyword expressions: "glob" for "glob"-style wildcard expressions;
"regex" for regular expressions; or "fixed" for exact matching. (copied from quanteda)}
}
\value{
data.frame with two columns: docname and the decision: "yer" or "not"
}
\description{
This function relies on the kwic function from package quanteda to mine the text of the judicial decisions
in order to identify whether it favored or not appelant.
}
\details{
.
}
\keyword{numbers}
