% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividir_sequencia.R
\name{dividir_sequencia}
\alias{dividir_sequencia}
\title{Divide sequência em n grupos}
\usage{
dividir_sequencia(sequencia, n)
}
\arguments{
\item{sequencia}{Sequência}

\item{n}{Número de grupos}
}
\value{
Lista
}
\description{
Divide sequência em n grupos
}
\examples{
dividir_sequencia(1:100, 5)
}
