% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editavel.R
\name{editavel}
\alias{editavel}
\title{Criar tabela para edição manual}
\usage{
editavel(x = NULL, coluna = NULL, acao = "     ",
  tamanho_pagina = 100, destino = "./editavel.html")
}
\arguments{
\item{x}{dataframe}

\item{coluna}{coluna de referência}

\item{acao}{coluna indicando acao  a ser tomada}

\item{tamanho_pagina}{padrão é cem, mas você pode aumentar.}

\item{destino}{diretorio/arquivo onde você quer salvar o html
editável. Se não indicar, será salvo um arquivo chamdo
editavel.html no atural diretorio.}
}
\value{
arquivo html que pode ser aberto por qualqueer navegador.
}
\description{
Criar tabela para edição manual
}
\details{
Esta função é especialmente importante quando queremos
    trabalhar em colaboração, pois ela gera um DT editável.
}
