% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminar_textos_indentados.R
\name{eliminar_textos_indentados}
\alias{eliminar_textos_indentados}
\title{Elimina as partes indentados de textos jurídicos}
\usage{
eliminar_textos_indentados(x)
}
\arguments{
\item{x}{Texto jurídico}
}
\value{
Mesmo texto sem partes indentadas
}
\description{
Elimina as partes indentados de textos jurídicos
}
