% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_ano.R
\name{extrair_ano}
\alias{extrair_ano}
\title{Extrai o ano do número do cnj}
\usage{
extrair_ano(processo, tamanho = c(25, 20))
}
\arguments{
\item{processo}{Número do processo}

\item{tamanho}{25, default, para números com ponto 
e hífein,  20, para números sem.}
}
\value{
Vetor número com pontos
}
\description{
Extrai o ano do número do cnj
}
