% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_numero_unico_pdf.R
\name{extrair_numero_unico_pdf}
\alias{extrair_numero_unico_pdf}
\title{Extrai número único do pdf}
\usage{
extrair_numero_unico_pdf(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos.}

\item{diretorio}{Caso não informe arquivos.}
}
\value{
Lista com números
}
\description{
Extrai número único do pdf
}
