% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_quantidade_droga.R
\name{extrair_quantidade_droga}
\alias{extrair_quantidade_droga}
\title{Extrair quantidade de droga da decisão sobre tráfico ou porte}
\usage{
extrair_quantidade_droga(x)
}
\arguments{
\item{x}{Texto}
}
\value{
Quantidade em gramas da droga
}
\description{
Extrair quantidade de droga da decisão sobre tráfico ou porte
}
\details{
Extrai somente quantidade em gramas
}
