% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_valor_maximo.R
\name{extrair_valor_arbitrado}
\alias{extrair_valor_arbitrado}
\title{Extrai valor máximo arbitrado}
\usage{
extrair_valor_arbitrado(julgado)
}
\arguments{
\item{julgado}{Inteiro teor do julgado}
}
\value{
Tibble
}
\description{
Extrai valor máximo arbitrado
}
