% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_search.R
\name{file_search}
\alias{file_search}
\title{Find text files that contains a pattern}
\usage{
file_search(pattern = "", dir = "R")
}
\arguments{
\item{pattern}{regex pattern}

\item{dir}{path}
}
\value{
Vector of tile names
}
\description{
Find text files that contains a pattern
}
\details{
Specially useful for changing dependencies inside R functions.
}
