% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrar_df.R
\name{filtrar_df}
\alias{filtrar_df}
\title{Filtra data.frame de textos com base em regex}
\usage{
filtrar_df(df = NULL, coluna = NULL, regex = NULL, n = 10)
}
\arguments{
\item{df}{Data.frame}

\item{coluna}{Decisões judiciais}

\item{regex}{Regex}

\item{n}{Quantidade de grupos para dividir o data.frame}
}
\value{
Mesmo data.frame filtrado
}
\description{
Filtra data.frame de textos com base em regex
}
\details{
Essa função é especialmente útil para dividir
    data.frames muito grandes em grupos antes de 
    apllicar o regex.
}
