% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_extrair.R
\name{gemini_extrair}
\alias{gemini_extrair}
\title{Extrai informações de textos judiciais com o GEMINI}
\usage{
gemini_extrair(
  x,
  api_key = NULL,
  instrucao,
  perguntas,
  colunas,
  candidate_count = 1,
  temperatura = 0,
  max_output_tokens = 4000,
  top_p = 0.4,
  top_k = 2
)
}
\arguments{
\item{x}{Texto}

\item{api_key}{Chave do GEMINI. Se não informada, buscara a variável de ambiente
GEMINI_API_KEY.}

\item{instrucao}{Orientação ao GEMINI, informando inclusive a delimitação por três
apóstrofes.}

\item{perguntas}{Vetor de perguntas.}

\item{colunas}{Vetor de colunas.}

\item{candidate_count}{Número de respostas}

\item{temperatura}{Nível de aleatoriedade. Padrão: 0, ou seja, determinístico.}

\item{max_output_tokens}{= Número máximo na resposta.}

\item{top_p}{Máxima probabilidade cumulativa para a amostra de tokens.}

\item{top_k}{Número máximo de tokens incluídos na amostra.}
}
\value{
json
}
\description{
Extrai informações de textos judiciais com o GEMINI
}
