% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_extrair.R
\name{gemini_extrair2}
\alias{gemini_extrair2}
\title{Extrai informações de textos judiciais com o GEMINI com saída estruturada}
\usage{
gemini_extrair2(
  x,
  api_key = NULL,
  instrucao,
  perguntas,
  colunas,
  modelo = "gemini-1.5-flash",
  temperatura = 1,
  max_output_tokens = 1e+05,
  top_p = 0.4,
  top_k = 2
)
}
\arguments{
\item{x}{Texto}

\item{api_key}{Chave do GEMINI. Se não informada, buscara a variável de ambiente
GEMINI_API_KEY.}

\item{instrucao}{Orientação ao GEMINI, informando inclusive a delimitação por três
apóstrofes.}

\item{perguntas}{Vetor de perguntas.}

\item{colunas}{lista  com json schema}

\item{modelo}{Modelo do Gemini}

\item{temperatura}{Nível de aleatoriedade. Padrão: 0, ou seja, determinístico.}

\item{max_output_tokens}{= Número máximo na resposta.}

\item{top_p}{Máxima probabilidade cumulativa para a amostra de tokens.}

\item{top_k}{Número máximo de tokens incluídos na amostra.}
}
\value{
json
}
\description{
Extrai informações de textos judiciais com o GEMINI com saída estruturada
}
