% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_son.R
\name{gpt_json}
\alias{gpt_json}
\title{Extrai informações do GPT}
\usage{
gpt_json(
  x,
  instrucao = NULL,
  perguntas,
  colunas,
  api_key = NULL,
  modelo = "gpt-4o-mini",
  temperatura = 0
)
}
\arguments{
\item{x}{Texto jurídico}

\item{instrucao}{Instrução para o GPT.}

\item{perguntas}{Vetor de perguntas.}

\item{colunas}{Lista com nomes das colunas do json
estruturada como no exemplo a seguir: list(resumo= list(type = "string"),
                                           merito = list(type = "string"
                                                          enum = c("procedente", "improcedente", "parcialmente procedente")),
                                           valor_condenacao = list(type = "number")
                                           )}

\item{api_key}{Chave OPENAI. Opcionalmente, armazene na como variável 
de ambiente: OPENAI_API_KEY}

\item{modelo}{Modelo da OPENAI. Padrão para gpt-4o-mini.}

\item{temperatura}{Temperatura. Padrão para 0.}
}
\value{
json
}
\description{
Extrai informações do GPT
}
