% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_ler.R
\name{gpt_ler}
\alias{gpt_ler}
\title{Lê respostas em arquivo json do GPT}
\usage{
gpt_ler(arquivos, colunas = NULL)
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{colunas}{Vetor de colunas a serem lidas}
}
\value{
tibble
}
\description{
Lê respostas em arquivo json do GPT
}
