% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt.R
\name{gpt_ler_cjpg}
\alias{gpt_ler_cjpg}
\title{Lê o json do GPT}
\usage{
gpt_ler_cjpg(arquivos, chaves = NULL)
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{chaves}{Vetor de chaves/colunas}
}
\value{
tibble
}
\description{
Lê o json do GPT
}
