% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperlink_excel.R
\name{hyperlink_excel}
\alias{hyperlink_excel}
\title{Mascara coluna com url(link) com hyperlink e salva em excel.}
\usage{
hyperlink_excel(data, col_link, col_texto, novo_nome = NULL, arquivo)
}
\arguments{
\item{data}{Dataframe com coluna com o link ou url a ser mascarada}

\item{col_link}{Coluna com link ou url}

\item{col_texto}{Coluna com texto (máscara)}

\item{novo_nome}{Novo nome da coluna. Opcional. Se não informada
receberá o mesmo nome da coluna col_link.}

\item{arquivo}{Nome do arquivo que armazenará o excel.}
}
\value{
arquivo excel
}
\description{
Mascara coluna com url(link) com hyperlink e salva em excel.
}
\examples{
\dontrun{

df <- data.frame(link = c("https://google.com","https://yahoo.com"),
                 nome = c("Google","Yahoo"))

hyperlink_excel(df, col_link = link, col_texto = nome, novo_nome = url, arquivo= "meu_excel.xlsx")

}
}
