% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inicial_final.R
\name{inicial_final}
\alias{inicial_final}
\title{Cria tibble com duas colunas de datas a partir para paralelização.}
\usage{
inicial_final(data_inicial = NULL, data_final = NULL,
  intervalos = 10, formato = "\%d/\%m/\%Y")
}
\arguments{
\item{data_inicial}{Data inicial no formato dd/mm/YYYY}

\item{data_final}{Data final no fomrato dd/mm/YYYY}

\item{intervalos}{Número de intervalos de datas criados}

\item{formato}{O padrão é dd/mm/YYYY, mas você pode especificar outro.}
}
\value{
tibble
}
\description{
Cria tibble com duas colunas de datas a partir para paralelização.
}
\examples{
\dontrun{
inicial_final(data_inicial="01/01/2017", data_final="31/12/2017")
}

}
