% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_son.R
\name{json_prompt}
\alias{json_prompt}
\title{Cria}
\usage{
json_prompt(x, instrucao, perguntas)
}
\arguments{
\item{x}{Texto jurídico (possivelmente)}

\item{instrucao}{Instrução}

\item{perguntas}{Vetor de perguntas}
}
\value{
lista com mensagens e instrução
}
\description{
Cria
}
