% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juris_wider.R
\name{juris_wider}
\alias{juris_wider}
\title{Converte para formato wide}
\usage{
juris_wider(df, valor, variavel)
}
\arguments{
\item{df}{tibble}

\item{valor}{valor}

\item{variavel}{variavel}
}
\value{
tibble
}
\description{
Converte para formato wide
}
