% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jurisMiner.R
\docType{package}
\name{JurisMiner-package}
\alias{JurisMiner}
\alias{JurisMiner-package}
\title{\code{jurisMiner} package}
\description{
Text mining of Brazilian judicial decisions
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/courtsbr/JurisMiner}
  \item Report bugs at \url{https://github.com/courtsbr/JurisMiner/issues}
}

}
\author{
\strong{Maintainer}: José de Jesus Filho \email{jjesus@consudata.org.br}

Other contributors:
\itemize{
  \item  Brazilian Jurimetrics Association [copyright holder]
}

}
