% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jus_kwic.R
\name{jus_kwic}
\alias{jus_kwic}
\title{Extrai contextos de padrões (regex)}
\usage{
jus_kwic(x, nomes = names(x), regex, compact = FALSE, tbl = FALSE)
}
\arguments{
\item{x}{Vetor de julgados}

\item{nomes}{Vetor com os nomes/números dos julgados.}

\item{regex}{Expressão regular}

\item{compact}{Remover nomes/números não encontrados?}

\item{tbl}{Converter resultado para tabela. Esta
opção remove não encontrados}
}
\value{
Lista ou tibble
}
\description{
Extrai contextos de padrões (regex)
}
