% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt.R
\name{jus_prompt}
\alias{jus_prompt}
\title{Auxilia na construção de mensagens}
\usage{
jus_prompt(x, destaques, perguntas, chaves)
}
\arguments{
\item{x}{decisão judicial}

\item{destaques}{Vetor de palavras ou expressões a destacar}

\item{perguntas}{Vetor de perguntas}

\item{chaves}{Vetor de chaves}
}
\value{
Lista de mensagens
}
\description{
Auxilia na construção de mensagens
}
