% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jus_unlist.R
\name{jus_unlist}
\alias{jus_unlist}
\title{Convert to vector list that contains empty or NULL}
\usage{
jus_unlist(list = NULL, default = NA_character_)
}
\arguments{
\item{list}{List containing empty values}

\item{default}{Default is NA_character_}
}
\value{
Vector with the same length of the list
}
\description{
Convert to vector list that contains empty or NULL
}
