% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_regex.R
\name{key_regex}
\alias{key_regex}
\title{Prepara um conjunto de palavras para busca com regex}
\usage{
key_regex(x = NULL)
}
\arguments{
\item{x}{Vetor de palavras ou expressões.}
}
\value{
string para ser buscada
}
\description{
Prepara um conjunto de palavras para busca com regex
}
