% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_pdfs.R
\name{ler_pdfs}
\alias{ler_pdfs}
\title{Wrapper de pdftools::pdf_text}
\usage{
ler_pdfs(arquivos = NULL, diretorio = ".", combinar = TRUE, basename = TRUE)
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Alternativamente indicar onde se encontram os arquivos}

\item{combinar}{Combinar as páginas num escalar? Padrão combinar}

\item{basename}{Cominho completo ou apenas o nome do arquivo. Padrão apenas nome.}
}
\value{
tibble
}
\description{
Wrapper de pdftools::pdf_text
}
