% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listar_arquivos.R
\name{listar_arquivos}
\alias{listar_arquivos}
\title{Lista arquivos rapidamente}
\usage{
listar_arquivos(diretorio = ".")
}
\arguments{
\item{diretorio}{Diretório onde se encontram os arquivos}
}
\value{
Vetor de arquivos com caminho absoluto
}
\description{
Lista arquivos rapidamente
}
\details{
Esta função tem o propósito de listar arquivos
    rapidamente, não ordenados. Não é recursiva, não aceita
    regex. Use list.files do base se forem poucos arquivos
    ou necessita de mais argumentos.
}
