% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_body.R
\name{make_body}
\alias{make_body}
\title{Cria lista para body de requisições http}
\usage{
make_body(query, dput = TRUE)
}
\arguments{
\item{query}{parsed query}

\item{dput}{TRUE para imprimir no console}
}
\value{
lista
}
\description{
Cria lista para body de requisições http
}
