% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_headers.R
\name{make_headers}
\alias{make_headers}
\title{Cria vetor de headers a partir de texto com headers}
\usage{
make_headers(x, cookies = FALSE)
}
\arguments{
\item{x}{String com headers}

\item{cookies}{Manter cookies? O padrão é remover.}
}
\value{
Vetor nomeado para ser usado em 
    httr::add_headers(.header= vetor)
}
\description{
Cria vetor de headers a partir de texto com headers
}
