% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_magistrados.R
\name{obter_magistrados}
\alias{obter_magistrados}
\title{Obter lista de magistrados}
\usage{
obter_magistrados(uf, justica = 1, grau = 1)
}
\arguments{
\item{uf}{Unidade Federativa, e.g., "SP"}

\item{justica}{Justica estadual (1) ou federal (2)}

\item{grau}{Somente primeiro grau funciona}
}
\value{
Dataframe com nomes de magistrados e
    tribunal
}
\description{
Obter lista de magistrados
}
