% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_contar_tokens.R
\name{openai_contar_tokens}
\alias{openai_contar_tokens}
\title{Conta tokens de textos com base em modelos da OPENAI}
\usage{
openai_contar_tokens(x, modelo = "gpt-3.5-turbo")
}
\arguments{
\item{x}{Vetor de textos}

\item{modelo}{Modelo a ser utilizado.}
}
\value{
Vetor com quantidade de tokens em cada texto.
}
\description{
Conta tokens de textos com base em modelos da OPENAI
}
\details{
Para usar esta função você tem de ter instalado o pacote
    tiktoken do Python, o qual será chamado via reticulate.
}
