% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pontua_cnj.R
\name{pontua_cnj}
\alias{pontua_cnj}
\title{Adiciona pontuação ao número do CNJ}
\usage{
pontua_cnj(numero = NULL)
}
\arguments{
\item{numero}{Número do cnj não pontuado. Por precaução,
esta função remove caracteres não numéricos e
adiciona zeros à esquerda caso não haja 20 dígitos.}
}
\value{
Número (character) devidamente pontuado
}
\description{
Adiciona pontuação ao número do CNJ
}
