% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pontua_cpf_cnpj.R
\name{pontua_cpf_cnpj}
\alias{pontua_cpf_cnpj}
\title{Pontua cnpj ou cpf}
\usage{
pontua_cpf_cnpj(x)
}
\arguments{
\item{x}{Vetor de cnpjs e ou cpfs.}
}
\value{
Mesmo vetor com valores pontuados
}
\description{
Pontua cnpj ou cpf
}
\examples{
pontua_cpf_cnpj(c("23268449000130","12345678901"))
}
