% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_kwic.R
\name{pt_kwic}
\alias{pt_kwic}
\title{Keyword in context for Portuguese texts}
\usage{
pt_kwic(string, docname = NULL, keyword = NULL, type = "coll",
  before = 5, after = 5, unite = FALSE)
}
\arguments{
\item{string}{a vector of texts from which to search for the keyword.}

\item{docname}{a vector of docnames. If ommited, it defaults
to text1,text2, text3...,}

\item{keyword}{keyword to search for}

\item{type}{the type of keyword pattern matching: "coll" for
collation, which will look for words, ignoring
case differences and diacritics; "regex" for regular
expressions. Regex will also ignore case differences and
diacritics. Option "coll" is much faster than "regex".}

\item{before}{Number of words before the keyword. Default is 5}

\item{after}{Number of words after the keyword. Default is 5}

\item{unite}{if FALSE, places every previous and posterior word in
separate column}
}
\value{
a tbl with docname, keyword location (start and end), the keyword,
    previous words, posterior words.
}
\description{
\code{pt_kwic} is an usefull function that allows you to extract
    words placed before and after a keyword. It is similar to 
    \code{\link[quanteda]{kwic}} from quanteda package, with two 
    important differences: it is dedicated to work only with
    Portuguese texts in a way that ignores diacritics and
    is not case sensitive; also it returns each word in a
    separate column as default.
}
\examples{
string<-c("A for\\u00e7a do direito deve superar o direito da for\\u00e7a.",
"Teu dever \\u00e9 lutar pelo Direito, mas se um dia encontrares o Direito
 em conflito com a Justi\\u00e7a,
luta pela Justi\\u00e7a.")
docname<-c("rui_barbosa","eduardo_couture")
keyword<-"direito"
df<-pt_kwic(string,docname,keyword)
}
