% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publicar_app.R
\name{publicar_app}
\alias{publicar_app}
\title{Envia shinyapp para srv/shiny-server}
\usage{
publicar_app(
  pacote = NULL,
  nome = NULL,
  destino = "/srv/shiny-server/",
  host = "localhost",
  port = 22,
  usuario = NULL,
  senha = NULL
)
}
\arguments{
\item{pacote}{Nome do pacote em que se encontra o shinyapp}

\item{nome}{Nome do diretório do shinyapp}

\item{destino}{Padrão /srv/shiny-server}

\item{host}{Se for localhost, envia para o servidor}

\item{port}{Se não for localhost, informar a porta ssh. Padrão 22.}

\item{usuario}{Usuário da máquina destino}

\item{senha}{Senha do usuário}
}
\description{
Envia shinyapp para srv/shiny-server
}
