% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparar_vetor.R
\name{reparar_vetor}
\alias{reparar_vetor}
\title{Prepara vetor de variáveis para aplicar juris_wider}
\usage{
reparar_vetor(x)
}
\arguments{
\item{x}{Vetor}
}
\value{
Vetor com variáveis em minúsculo, sem acentos
    e espaços e pontuação substituídos por _.
}
\description{
Prepara vetor de variáveis para aplicar juris_wider
}
