% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumir_inteiro_teor.R
\name{resumir_inteiro_teor}
\alias{resumir_inteiro_teor}
\title{Resume inteiro teor de decisões judiciais}
\usage{
resumir_inteiro_teor(x, model = "gpt-3.5-turbo-16k", temperatura = 0)
}
\arguments{
\item{x}{Vetor de decisões}

\item{model}{Modelo}

\item{temperatura}{Grau de aleatoriedade. Padrão zero.}
}
\value{
resumo da decisão
}
\description{
Resume inteiro teor de decisões judiciais
}
