% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_stopwords.R
\name{rm_stopwords}
\alias{rm_stopwords}
\title{Remove Brazilian stopwords from strings}
\usage{
rm_stopwords(string, stopwords)
}
\arguments{
\item{string}{character vector of strings}

\item{stopwords}{word to be removed}
}
\value{
Same object without stopwords
}
\description{
Remove Brazilian stopwords from strings
}
