% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg_decison.R
\name{sg_decision}
\alias{sg_decision}
\title{Search for second instance judicial decision}
\usage{
sg_decision(string, docname, keyword, class = "appeal")
}
\arguments{
\item{string}{a vector of texts from which to search for the keyword.}

\item{docname}{a vector of docnames. If ommited, it defaults
to text1,text2, text3...,}

\item{keyword}{keyword to search for. For appeals, e.g, apelação and agravo,
 the keyword is "provimento". For writs, e.g, habeas corpus and "mandado
"segurança", the keyword should be "concedida" or "concedido", or just a
regex: "concedid*"}

\item{class}{procedural class, either "appeal" or "writ".}
}
\value{
a tibble with two columns: docname and decision: c("sim","não","parcial")
}
\description{
Search for second instance judicial decision
}
\examples{
\dontrun{
df<-sg_decision(string,docname,"provimento")
}
}
