% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakecase.R
\name{snakecase}
\alias{snakecase}
\title{Wrapper para snakecase sem acentuação.}
\usage{
snakecase(x)
}
\arguments{
\item{x}{Converte para snakecase sem acentos.}
}
\value{
x em snakecase
}
\description{
Wrapper para snakecase sem acentuação.
}
