% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{Common Brazilian stopwords}
\format{string with stopwords}
\usage{
stopwords
}
\description{
String with common stopwords
}
\details{
\itemize{
  \item string with stopwords
  ...
}
}
\keyword{datasets}
