% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempo_movimentacao.R
\name{tempo_movimentacao}
\alias{tempo_movimentacao}
\title{Calcula tempo entre movimentações processuais}
\usage{
tempo_movimentacao(df, data = data)
}
\arguments{
\item{df}{Tibble com movimentação processual}

\item{data}{Coluna (bear) com datas no formato
iso 8601}
}
\value{
Mesma tibble adicionada das colunas
    decorrencia e decorrencia acumulada
}
\description{
Calcula tempo entre movimentações processuais
}
\examples{
\dontrun{
df <- tempo_movimentacao(df)
}
}
