% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpu_detalhes_item_publico.R
\name{tpu_detalhes_item_publico}
\alias{tpu_detalhes_item_publico}
\title{Pesquisa ítem público das tabelas processuais unificadas}
\usage{
tpu_detalhes_item_publico(tipo_tabela, cod_item)
}
\arguments{
\item{tipo_tabela}{Tipo da tabela a ser pesquisada(A,M,C) - Assuntos, Movimentos, Classes}

\item{cod_item}{Tipo da pesquisa(G,N,C) - Glossário, Nome, Código}
}
\value{
Tibble com colunas chave e valor. Além disso, argumentos de 
    pesquisa estão associados. Chame `str(df)` para ver os atributos.
}
\description{
Pesquisa ítem público das tabelas processuais unificadas
}
\examples{
\dontrun{
 # Pesquisa classe cujo código é 156
 df <- tpu_detalhes_item_publico(
                          tipo_tabela = "C",
                          cod_item = "156"
                          ) 
}
}
