% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpu_extrair_ramo.R
\name{tpu_extrai_ramo}
\alias{tpu_extrai_ramo}
\title{Extrair ramo do assunto}
\usage{
tpu_extrai_ramo(cod_item)
}
\arguments{
\item{cod_item}{Código do assunto}
}
\value{
Codigo do ramo e ramo
}
\description{
Extrair ramo do assunto
}
