% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpu_get_ultima_versao.R
\name{tpu_get_ultima_versao}
\alias{tpu_get_ultima_versao}
\title{Verifica última versão}
\usage{
tpu_get_ultima_versao()
}
\value{
Data com classe Date
}
\description{
Verifica última versão
}
\examples{
\dontrun{
 # Pesquisa classe cujo código é 156
 tpu_pesquisar_item_publico(
                          tipo_tabela = "C",
                          tipo_pesquisa = "C",
                          valor_pesquisa = "156"
                          ) 
}
}
