% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpu_pesquisar_item_publico.R
\name{tpu_pesquisar_item_publico}
\alias{tpu_pesquisar_item_publico}
\title{Pesquisa ítem público das tabelas processuais unificadas}
\usage{
tpu_pesquisar_item_publico(tipo_tabela, tipo_pesquisa, valor_pesquisa)
}
\arguments{
\item{tipo_tabela}{Tipo da tabela a ser pesquisada(A,M,C) - Assuntos, Movimentos, Classes}

\item{tipo_pesquisa}{Tipo da pesquisa(G,N,C) - Glossário, Nome, Código}

\item{valor_pesquisa}{Valor da pesquisa}
}
\value{
Tibble
}
\description{
Pesquisa ítem público das tabelas processuais unificadas
}
\examples{
\dontrun{
 # Pesquisa classe cujo código é 156
 tpu_pesquisar_item_publico(
                          tipo_tabela = "C",
                          tipo_pesquisa = "C",
                          valor_pesquisa = "156"
                          ) 
}
}
