% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearPt.R
\name{clearPt}
\alias{clearPt}
\title{Optionally removes stopwords, numbers, ordinals, whitespaces, punctuation, diacritics, and convert to lower case.}
\usage{
clearPt(string, lower = TRUE, stopwords = TRUE, diacritics = TRUE,
  punctuation = TRUE, squish = TRUE, numbers = TRUE,
  ordinal = TRUE)
}
\arguments{
\item{string}{object to be submitted to conversion and removals.}

\item{lower}{logical Converts to lower case.}

\item{stopwords}{logial Removes portuguese-BR stopwords.}

\item{diacritics}{logial remove all diacritics}

\item{punctuation}{logical Romoves punctuation.}

\item{squish}{logical applies stringr::str_squish}

\item{numbers}{logical Removes numbers}

\item{ordinal}{logical Removes ordinal superscripts}
}
\description{
Optionally removes stopwords, numbers, ordinals, whitespaces, punctuation, diacritics, and convert to lower case.
}
\examples{
\dontrun{
clearPt(texto)
}
}
