% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearPt.R
\name{clearPt}
\alias{clearPt}
\title{Function clearPt}
\usage{
clearPt(x, lower = T, stopwords = T, accent = T, punctuation = T,
  whitespace = T, numbers = T, ordinal = T)
}
\arguments{
\item{x}{object to submitted to translation and removals.}

\item{lower}{logical: Translate to lower case.}

\item{stopwords}{logial: Removes portuguese-BR stopwords.}

\item{accent}{logial: Encode to ASCII so that it removes all accents.}

\item{punctuation}{logical: Romoves punctuation.}

\item{whitespace}{logical: Remove whitespaces}

\item{numbers}{logical: Removes numbers}

\item{ordinal}{logical: Remover indicadores de ordinais}
}
\description{
Optionally Removes stopwords, numbers, ordinals, whitespaces, punctuation, accents, and translate to lower case.
}
\examples{
clearPt(texto)
}
\keyword{accent,}
\keyword{clean,}
\keyword{lower.}
\keyword{punctuation,}
\keyword{stopwords,}
\keyword{tranlation,}
