% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnj_quantidade.R
\name{cnj_quantidade}
\alias{cnj_quantidade}
\title{Encontra a quantidade de processos distribuídos numa unidade judiciária.}
\usage{
cnj_quantidade(fim, inicio = 0, ano, nivel, uf, distribuidor, funcao,
  expr = "sp_vazio")
}
\arguments{
\item{fim}{Qualquer inteiro que seguramente é superior ao número de processos distribuídos}

\item{ano}{Indicar o ano em questão}

\item{nivel}{Nível do judiciário}

\item{uf}{Unidade federativa.}

\item{distribuidor}{Código do distribuídor. Encontrar no data-raw.}

\item{funcao}{Função a ser aplicada para baixar: baixar_cpopg, baixar_processo, etc.}

\item{expr}{Expressão a ser avaliada. No caso de São Paulo, eu criei uma função
interna chamada `sp_vazio`, que verifica se o tamanho do arquivo é menor que 90Mb.}
}
\value{
Quantidade máxima aproximada de processos distribuídos. Pode haver um pequeno erro
}
\description{
Encontra a quantidade de processos distribuídos numa unidade judiciária.
}
