% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnj_ultimo.R
\name{cnj_ultimo}
\alias{cnj_ultimo}
\title{Encontra a quantidade de processos distribuídos numa unidade judiciária.}
\usage{
cnj_ultimo(x, ano, nivel, uf, distribuidor, funcao, expr = "is.null")
}
\arguments{
\item{x}{Qualquer inteiro que seguramente é superior ao número de processos distribuídos}

\item{ano}{Indicar o ano em questão}

\item{nivel}{Nível do judiciário}

\item{uf}{Unidade federativa.}

\item{distribuidor}{Código do distribuídor. Encontrar no data-raw.}

\item{funcao}{Função a ser aplicada para baixar: baixar_cpopg, baixar_processo, etc.}

\item{expr}{Expressão a ser avaliada. A expressão deve retornar TRUE para processos
não encontrados. Em São Paulo, o tamanho deve ser menor que 86K, no DF, a resposta NULL.}
}
\value{
Quantidade máxima aproximada de processos distribuídos. Pode haver um pequeno erro
}
\description{
Encontra a quantidade de processos distribuídos numa unidade judiciária.
}
