% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_ano.R
\name{extrair_ano}
\alias{extrair_ano}
\title{Extrai o ano do número do cnj}
\usage{
extrair_ano(processo)
}
\arguments{
\item{processo}{Número do processo}
}
\value{
Vetor número com pontos
}
\description{
Extrai o ano do número do cnj
}
