% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_kwic.R
\name{pt_kwic}
\alias{pt_kwic}
\title{Keyword in context for Portuguese texts}
\usage{
pt_kwic(
  string,
  id_decision = NULL,
  keyword = NULL,
  before = 5,
  after = 5,
  unite = TRUE
)
}
\arguments{
\item{string}{a vector of texts from which to search for the keyword.}

\item{id_decision}{a vector of id_decisions. If ommited, it defaults
to text1,text2, text3...,}

\item{keyword}{you can provide a regex expression or a vector of regex
expressions.}

\item{before}{Number of words before the keyword. Default is 5}

\item{after}{Number of words after the keyword. Default is 5}

\item{unite}{if FALSE, places every previous and posterior word in
separate column}
}
\value{
a tbl with id_decision, keyword location (start and end), the keyword,
    the previous words, and the posterior words.
}
\description{
\code{pt_kwic} is an useful function that allows you to extract
    words placed before and after a keyword. It is similar to 
    \code{\link[quanteda]{kwic}} from quanteda package, with two 
    important differences: it is dedicated to work only with
    Portuguese texts in a way that ignores diacritics and
    is case insensitive; also it returns each word in a
    separate column as a default.
}
\examples{
string<-c("A força do direito deve 
superar o direito da força.",
"Teu dever é lutar pelo Direito,
 mas se um dia encontrares o Direito
 em conflito com a Justiça,
luta pela Justiça.")
id_decision<-c("rui_barbosa","eduardo_couture")
keyword<-"direito"
df<-pt_kwic(string,id_decision,keyword)
}
