% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_stopwords.R
\name{rm_stopwords}
\alias{rm_stopwords}
\title{Remove stopwords de decisões judiciais}
\usage{
rm_stopwords(texto, stopwords)
}
\arguments{
\item{texto}{}

\item{stopwords}{}
}
\value{
Mesmo objeto texto com stopwords removidas
}
\description{
Remove stopwords de decisões judiciais
}
\examples{

}
