% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg_decison.R
\name{sg_decision}
\alias{sg_decision}
\title{Classifies high courts opinions}
\usage{
sg_decision(string, id_decision, keyword, class = "appeal",
  type = "coll")
}
\arguments{
\item{string}{a vector of texts from which to search for the keyword.}

\item{id_decision}{a vector of id_decisions. If ommited, it defaults
to text1,text2, text3...,}

\item{keyword}{keyword to search for. For appeals, e.g, apelação and agravo,
 the keyword is usually "provimento". For writs, e.g, habeas corpus and "mandado
"segurança", the keyword should be "concedida" or "concedido", or just a
regex: "concedid*"}

\item{class}{procedural class, either "appeal" (default) or "writ".}

\item{type}{It can be either "coll" for collation search or "regex".}
}
\value{
a tbl with the id_decision, the keyword location (start and end), the keyword,
    the previous words, the posterior words, and the decision: c("sim","não","parcial")
}
\description{
This function will carry out a search for a keyword such as "affirm" or "reverse",
  but in Portuguese, from hight courts opinions already read to R and do a first attempt 
  to classify them
}
\examples{
\dontrun{
df<-sg_decision(string,id_decision,"provimento")
}
}
